// local-hobbies-frontend/src/types/availability.ts
export enum AvailabilityType {
    OneTimeAvailability = 'One-time availability',
    RecurringAvailability = 'Recurring availability',
    AvailabilityException = 'Availability exception',
}

export enum AvailabilityFrequency {
    Weekly = 'Weekly',
    EveryTwoWeeks = 'Every two weeks',
    Monthly = 'Monthly',
}

export enum DayOfWeek {
    Monday = 'Monday',
    Tuesday = 'Tuesday',
    Wednesday = 'Wednesday',
    Thursday = 'Thursday',
    Friday = 'Friday',
    Saturday = 'Saturday',
    Sunday = 'Sunday',
}
// ...
export interface OneTimeAvailabilityResponse {
    id: string;
    location: GeoJsonPoint;
    /** ISO 8601 UTC timestamp */
    start: string;
    /** ISO 8601 duration */
    duration: string;
}
// ...
export interface RecurringAvailabilityResponse {
    id: string;
    location: GeoJsonPoint;
    /** ISO 8601 date */
    ruleStart: string;
    /** ISO 8601 date */
    ruleEnd: string | null;
    frequency: AvailabilityFrequency;
    startDayOfWeek: DayOfWeek | null;
    startDayOfMonth: number | null;
    /** ISO 8601 time */
    startTime: string;
    /** ISO 8601 duration */
    duration: string;
}
// ...
export interface AvailabilityExceptionResponse {
    id: string;
    recurringAvailabilityId: string;
    /** ISO 8601 date */
    exceptionDate: string;
    exceptionReason: string;
    isCancelled: boolean;
    overrideLocation: GeoJsonPoint | null;
    /** ISO 8601 time */
    overrideStartTime: string | null;
    /** ISO 8601 duration */
    overrideDuration: string | null;
}
// ...
export interface ScheduleResponse {
    intervals: AvailabilityIntervalResponse[];
    availabilities: {
        oneTimes: OneTimeAvailabilityResponse[];
        recurrings: RecurringAvailabilityResponse[];
        exceptions: AvailabilityExceptionResponse[];
    };
}

export interface AvailabilityIntervalResponse {
    sourceType: AvailabilityType;
    sourceId: string;
    location: GeoJsonPoint;
    /** ISO 8601 UTC timestamp */
    start: string;
    /** ISO 8601 UTC timestamp */
    end: string;
}